requirements.txt
pandas==2.0.1
pyarrow==11.0.0
seaborn==0.12.2
scikit-learn==1.2.2
mlflow==2.3.2
hyperopt==0.2.7
xgboost==1.7.5
optuna==3.1.1
wandb==0.15.3
black==23.3.0
prefect==2.10.8
prefect-aws==0.3.1
orjson==3.8.1
python-dotenv==1.0.0
